% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_niche_metadata.R
\name{add_niche_metadata}
\alias{add_niche_metadata}
\title{Add niche summary information to Seurat metadata}
\usage{
add_niche_metadata(obj, niche_summary)
}
\arguments{
\item{obj}{A Seurat object.}

\item{niche_summary}{A data.frame returned by \code{summarize_niches()} with (at minimum) Niche and Niche_Type columns.}
}
\value{
Seurat object with added metadata columns.
}
\description{
Merge niche-level summary information back into cell-level metadata and add to the Seurat object.
}
\examples{
\dontrun{
obj <- add_niche_metadata(obj, niche_summary)
}

}
