% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_format.R
\name{data_format}
\alias{data_format}
\title{Convert SpatialExperiment to Seurat}
\usage{
data_format(obj)
}
\arguments{
\item{obj}{An R object. Expected to be a \code{Seurat} or \code{SpatialExperiment}.}
}
\value{
A \code{Seurat} object.
}
\description{
Convert a \code{Seurat} or \code{SpatialExperiment} object to a Seurat object.
If the input is already a \code{Seurat} object, it is returned unchanged.
}
\examples{
\dontrun{
seurat_obj <- data_format(my_spatial_experiment)
}

}
