% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_niche_plot.R
\name{make_niche_plot}
\alias{make_niche_plot}
\title{Plot TLS niche overlay on spatial image}
\usage{
make_niche_plot(
  obj,
  structure_label = "TLS niche",
  tissue_overlay = FALSE,
  structure_color = "#D81B60"
)
}
\arguments{
\item{obj}{Seurat object with \code{Niche_Type} metadata.}

\item{structure_label}{Character. Label used to identify the TLS-rich niche. Default "TLS niche".}

\item{tissue_overlay}{Logical. If TRUE, use SpatialDimPlot (requires histology).}

\item{structure_color}{Character. Color of structure_label on spatial image. Default "#D81B60".}
}
\value{
A ggplot object.
}
\description{
Create a spatial overlay that highlights TLS candidate niches vs other niches.
}
\examples{
\dontrun{
niche_image <- make_niche_plot(obj, 
structure_label = "TLS niche", 
tissue_overlay = TRUE, 
structure_color = "blue")
}

}
