% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_spatial_data.R
\name{process_spatial_data}
\alias{process_spatial_data}
\title{Normalize and scale spatial transcriptomic data}
\usage{
process_spatial_data(seurat_obj, assay, method = c("sctransform", "lognorm"))
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{assay}{Character. Name of the assay to normalize (e.g., "Xenium" or "SCT").}

\item{method}{Character. One of "sctransform" or "lognorm". Default "sctransform".}
}
\value{
The normalized and scaled Seurat object.
}
\description{
Normalize and scale a Seurat object using either SCTransform or log-normalize + FindVariableFeatures + ScaleData.
}
\examples{
\dontrun{
obj <- process_spatial_data(seurat_obj, assay = "Xenium", method = "sctransform")
}

}
