% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_annotation.R
\name{run_annotation}
\alias{run_annotation}
\title{Annotate cells using SingleR and celldex reference}
\usage{
run_annotation(
  seurat_obj,
  reference = "HumanPrimaryCellAtlasData",
  assay = c("Xenium", "SCT"),
  labels_name = "SingleR_label"
)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{reference}{Character. One of "HumanPrimaryCellAtlasData", "BlueprintEncodeData", "ImmGenData", "MouseRNAseqData", "MonacoImmuneData".}

\item{assay}{Character or name. The assay to pull expression from (e.g., "Xenium", "SCT").}

\item{labels_name}{Character. Metadata column name to hold assigned labels. Default "SingleR_label".}
}
\value{
Seurat object with new metadata column containing SingleR labels.
}
\description{
Run SingleR annotation on expression data extracted from a spatial Seurat object using a celldex reference dataset.
}
\examples{
\dontrun{
obj <- run_annotation(seurat_obj, reference = "HumanPrimaryCellAtlasData", assay = "SCT")
}

}
